Remove unneeded compatibility shim
