+++ phpcs.xml.dist
@@ -115,8 +115,6 @@
 	<!-- Exclude existing classes from the protected parameter names rule as it would break BC to change these now. -->
 	<rule ref="Universal.NamingConventions.NoReservedKeywordParameterNames">
 		<exclude-pattern>*/php/WP_CLI/Iterators/Transform\.php</exclude-pattern>
-		<exclude-pattern>*/php/WP_CLI/Compat/Min_PHP_5_4/FeedbackMethodTrait\.php</exclude-pattern>
-		<exclude-pattern>*/php/WP_CLI/Compat/Min_PHP_5_6/FeedbackMethodTrait\.php</exclude-pattern>
 		<exclude-pattern>*/php/WP_CLI/Inflector\.php</exclude-pattern>
 		<exclude-pattern>*/php/WP_CLI/Autoloader\.php</exclude-pattern>
 		<exclude-pattern>*/php/WP_CLI/Dispatcher/CommandFactory\.php</exclude-pattern>
